Fix Docker tag in make release target
Fix Docker tag in make release target so that it is NEXTTAG, the value
with which we tag source code and github release. If the last release
was v0.0.1, NEXTTAG will be v0.0.2; Until now the docker tag would have
been v0.0.1-gabcd, where abcd is the newly added sha.